requirements.txt
asgiref==3.2.7
Django==3.0.6
djangorestframework==3.11.0
pkg-resources==0.0.0
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
tqdm==4.46.0
